refactor(backend): extend Rust route workflows and runtime services
- update AI clients and route handlers for expanded backend coverage
- split chapter and project flows into more focused runtime services
- align settings, memory, MCP, and export paths with the new workflow model

+++ backend-rs/src/tasks/types.rs
@@ -1,7 +1,7 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
 pub enum TaskStatus {
     Pending,
@@ -117,7 +117,7 @@ pub struct TaskListQuery {
     pub project_id: Option<String>,
     pub statuses: Option<String>,
     pub active_only: Option<bool>,
-    pub limit: Option<usize>,
+    pub limit: Option<i64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
